revert mock data ownership
