+++ test/BaseTest.ts
@@ -1,7 +1,7 @@
 //@ts-nocheck
-import { vi, expect, test, describe } from "vitest";
+import { vi, expect, test as testBase, describe } from "vitest";
 import puppeteer from 'puppeteer';
-import type { Browser, Page } from "puppeteer";
+
 import { BasemapSession, BasemapStyle } from '../src/MaplibreArcGIS.js';
 import { ApiKeyManager } from '@esri/arcgis-rest-request';
 import { Map } from 'maplibre-gl';
@@ -9,8 +9,50 @@ import {MOCK_API_KEY} from './mock/authentication/basemapApiKey.js';
 import basemapStyleNavigation from './mock/BasemapStyle/ArcGISNavigation.json';
 import sessionResponseRaw from './mock/BasemapSession/valid-session.json';
 
+// Mock service containing multiple feature layers (12 layers)
+import multiLayerServiceDefinitionRaw from './mock/FeatureLayer/multiLayer-service-info.json';
+
+const featureMultiLayerMock = {
+  itemId: '44299709cce447ea99014ff1e3bf8505',
+  serviceUrl: 'https://services3.arcgis.com/GVgbJbqm8hXASVYi/arcgis/rest/services/ManyLayers/FeatureServer',
+  serviceDefinition: JSON.stringify(multiLayerServiceDefinitionRaw),
+  serviceDefinitionRaw: multiLayerServiceDefinitionRaw
+}
+
+// Santa Monica trails feature mocks
+import trailsLayerDefinitionRaw from './mock/FeatureLayer/trails/trails-layer-info.json';
+import trailsServiceDefinitionRaw from './mock/FeatureLayer/trails/trails-service-info.json';
+import trailsItemRaw from './mock/FeatureLayer/trails/trails-item-info.json';
+import trailsDataRaw from './mock/FeatureLayer/trails/trails-features.json';
+import trailsDataTruncatedRaw from './mock/FeatureLayer/trails/trails-features-truncated.json'
+import trailsQueryDataRaw from './mock/FeatureLayer/trails/trails-feature-query.json';
+import trailsExceedsLimitRaw from './mock/FeatureLayer/trails/trails-feature-exceedsLimit.json';
+
+const featureTrailsMock = {
+  itemId: '69e12682738e467eb509d8b54dc73cbd',
+  item: JSON.stringify(trailsItemRaw),
+  itemRaw: trailsItemRaw,
+  serviceUrl: 'https://services3.arcgis.com/GVgbJbqm8hXASVYi/arcgis/rest/services/Trails/FeatureServer',
+  serviceDefinition: JSON.stringify(trailsServiceDefinitionRaw),
+  serviceDefinitionRaw: trailsServiceDefinitionRaw,
+  layerUrl: 'https://services3.arcgis.com/GVgbJbqm8hXASVYi/arcgis/rest/services/Trails/FeatureServer/0',
+  layerDefinition: JSON.stringify(trailsLayerDefinitionRaw),
+  layerDefinitionRaw: trailsLayerDefinitionRaw,
+  exceedsLimitResponse: JSON.stringify(trailsExceedsLimitRaw),
+  exceedsLimitResponseRaw: trailsExceedsLimitRaw,
+  geoJSONSmall: JSON.stringify(trailsDataTruncatedRaw),
+  geoJSONLarge: JSON.stringify(trailsDataRaw),
+  geoJSONRaw: trailsDataRaw,
+  geoJSONSmallRaw: trailsDataTruncatedRaw
+};
+
+export const featureMocks = {
+  trailsMock: featureTrailsMock,
+  multiLayerMock: featureMultiLayerMock
+};
+
 let browser;
-async function setupBrowser(): Promise<Browser> {
+async function setupBrowser() {
   if (browser) {
     return browser
   };
@@ -21,7 +63,7 @@ async function setupBrowser(): Promise<Browser> {
   return browser;
 }
 
-export const customTest = test.extend({
+export const customTest = testBase.extend({
   // API key
   apiKey: async ({}, use) => {
     await use (process.env.PRODUCTION_KEY_ALP || "fake-api-key");
@@ -76,7 +118,7 @@ export const customTest = test.extend({
     await use(basemap);
   },
   setupPage : async ({}, use) => {
-    async function loadPage (mockPageFile): Promise<Page> {
+    async function loadPage (mockPageFile) {
       //setup the browser if it isn't already
       const browser = await setupBrowser();
