+++ test/mock/pages/feature-layer.html
@@ -13,29 +13,21 @@
     }
   </style>
   <link href="../../../node_modules/maplibre-gl/dist/maplibre-gl.css" rel="stylesheet" />
+  <script src="../../../node_modules/maplibre-gl/dist/maplibre-gl.js"></script>
+  <script src="../../../dist/umd/maplibre-arcgis.min.js"></script>
 </head>
 
 <body>
   <div id="map"></div>
-  <script type="importmap">
-    {
-      "imports": {
-        "maplibre-gl": "../../../node_modules/maplibre-gl/dist/maplibre-gl.mjs",
-        "maplibre-arcgis": "../../../dist/esm/maplibre-arcgis.min.js"
-      }
-    }
-  </script>
   <script type="module">
-    import { Map } from "maplibre-gl"
-    import { FeatureLayer } from "maplibre-arcgis"
-    const map = new Map({
+    const map = new maplibregl.Map({
       container: "map",
       zoom: 5,
       center: [-118.4882,34.0243]
     });
 
     const layerUrlTrails = 'https://services3.arcgis.com/GVgbJbqm8hXASVYi/arcgis/rest/services/Trails/FeatureServer/0';
-    const featureLayer = new FeatureLayer({
+    const featureLayer = new maplibreArcGIS.FeatureLayer({
       url:layerUrlTrails,
       loadingMode:'ondemand'
     });
